requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
numpy>=1.26.0
opencv-python-headless>=4.9.0
pillow>=10.2.0
deepface>=0.0.87
